[Backport 2.20][PLAT-9084] Support LDAP with RBAC
Summary:
Today, we fetch the role from LDAP or the LDAP group mapping and fall back to default if a user is not identified.
Based on the role identified in the above step, corresponding role bindings are generated.
The scope of these roles is set to all the resources under the customer to which the user belongs.
The roles that can be specified in the LDAP configuration should be System-defined ones; the custom ones are not currently supported.

Made a change not to perform any authorization using the TokenAuthenticator in case the runtime config for the new RBAC is not set to true.

Original commit: https://github.com/yugabyte/yugabyte-db/commit/55afdde751142ffa76a01f195c532d591ae75259
Original PR: https://phorge.dev.yugabyte.com/D28663

Test Plan:
Test the changes by setting up the LDAP server and running the LDAP server with various users
Verified that the corresponding role bindings are getting populated in the DB
Tested that the authorization of this user is working as expected.

UTs

Reviewers: #yba-api-review, sneelakantan, skurapati

Reviewed By: #yba-api-review, sneelakantan

Subscribers: yugaware, skurapati

Differential Revision: https://phorge.dev.yugabyte.com/D28957