[BACKPORT 2.20][PLAT-8975] During ysql upgrade, select active tserver in same region as master leader
Summary:
Original commit:
2a15be36747e4ebdb3035f81a14f0fa2c44d4e43 / D28594

In multi-region universe set up. When performing ysql upgrade (subtask), the upgradeYSQL rpc loops through every database and tries to get a postgres connection. This may take a while if the tserver we are using for the ysql upgrade is far away from the master leader (i.e. different regions). Multiple calls with this adds up to a significant amount of time.

Optimized this by selecting an active tserver in the same region as master leader if possible. If not, we will select a random active tserver from any region.

Test Plan:
Create a multi-region 3 node rf3 universe. Regions: us-west1, us-west2, us-east1.

Perform a software upgrade. After the universe has just succeeded, check the region the master leader is in. Say the master leader is in us-west2. Check the logs and make sure that the ysqlUpgrade chose the tserver in the us-west2 region.

Example:
```Tserver: yb-admin-cwang-multi-region-check-n3 is alive in region us-west2, with ip and port: 10.158.0.13:9100```

Reviewers: hzare, nsingh, sanketh

Reviewed By: hzare

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D28818