[BACKPORT 2.20][#19394] CDCSDK: Set snapshot key from correct parameter in explicit checkpointing
Summary:
Original commit: bc343e7b7882fb88aec833814076868a8eea398c / D29060

This diff modifies a method in `AsyncYBClient` which was earlier passing a `null` value to another overloaded method. The effect of this was that the `explicit_cdc_sdk_opid` was not even being passed to the service.

Another thing this diff modifies is that while setting the checkpoint during snapshot, service was using the `from_op_id` whereas we should be using the `explicit_cdc_sdk_opid` if the stream is in `EXPLICIT` checkpointing mode.

**2.20 only note:**
Skipping the `yb-client` version changes from the original diff while backporting.
Jira: DB-8188

Test Plan: Run existing tests to verify no regression.

Reviewers: skumar, stiwary, asrinivasan

Reviewed By: asrinivasan

Subscribers: ycdcxcluster

Differential Revision: https://phorge.dev.yugabyte.com/D29122