[BACKPORT 2.20][#19394] CDCSDK: Only populate key to `GetChangesRequest` when it is not null
Summary:
Original commit: 64642e2b0d59c38a19f26e08a8c4ba6fd2c617a2 / D29124

This diff modifies the way we populate the key in the `from_op_id` and `explicit_op_id` parameter of `GetChangesRequest` and now we will only set the key if it is not `null` - if it is `null` then we will not do anything to key.

**2.20 only note:**
Skipping the version changes in yb-client.
Jira: DB-8188

Test Plan: Run existing tests.

Reviewers: skumar, stiwary, asrinivasan

Reviewed By: skumar

Subscribers: ycdcxcluster

Differential Revision: https://phorge.dev.yugabyte.com/D29193