[BACKPORT 2.20][#19434] CDCSDK: Do not completely fail while fetching tablets if one table hits error
Summary:
Original commit: 6875975cc7f51c80eb6086e0ad7f1adf04faacee / D29189

While fetching all the tablets for a given table, we return with status code not okay if any table in the list hits an error. This doesn't allow the flow to move forward to fetch tablets for other tables for which streaming is fine and the service ends up sending unnecessary errors to the client.

This diff modifies the logic to simply log a warning if something fails while fetching tablets for a given table and continue. Note that the flow affected by this diff is just in the method `GetChanges`.
Jira: DB-8227

Test Plan:
Jenkins: skip

```
ybd --cxx-test cdcsdk_ysql-test --gtest_filter CDCSDKYsqlTest.TestUnrelatedTableDropUponTserverRestart
```

Reviewers: skumar, asrinivasan, stiwary

Reviewed By: skumar

Subscribers: ycdcxcluster

Tags: #jenkins-ready

Differential Revision: https://phorge.dev.yugabyte.com/D29369