[BACKPORT 2.20][#19197] DocDB: Fixes flaky test PgGetLockStatusTest.TestLocksOfColocatedTables
Summary:
Original commit: fcaf5ff56b0215c6458ac70e9d1ca58519802be6 / D29066
Prior to this diff, we seem to not wait for creation of local transactions table in `PgLocksTestBase` which extends `GeoTransactionsTestBase`. This seems to be the reason for the flakiness of the all colocation tests of `PgGetLockStatusTest` in release mode.
Jira: DB-7994

Test Plan:
Jenkins

./yb_build.sh release  --cxx-test pgwrapper_pg_get_lock_status-test --gtest_filter PgGetLockStatusTest.TestLocksOfColocatedTables -n 10 --tp 1
./yb_build.sh release  --cxx-test pgwrapper_pg_get_lock_status-test --gtest_filter PgGetLockStatusTest.TestColocatedWaiterWriteLock -n 10 --tp 1

Reviewers: rsami, rthallam

Reviewed By: rthallam

Subscribers: yql, ybase, bogdan

Tags: #jenkins-ready

Differential Revision: https://phorge.dev.yugabyte.com/D29440