[BACKPORT 2.20][#13358] YSQL: YSQL DDL Atomicity Part 5: YB-Backup support
Summary:
Original commit: 20b330695d58fa4840ab558e1c1134ec74f97721 / D27966
**Problem**
This patch handles the case where YB-Backup is performed when Ddl verification
is still underway. Currently there is a check in the yb_backup.py script that
loops on the catalog version to verify that no Ddls have happened during the
backupprocess. However, if it is only the DDL verification that happened after a
successful/failed DDL then this will not be picked up the catalog version check.

Thus, it is possible for backup to have run when the PG catalog and DocDB
schema were not in sync with each other. This could lead to the snapshot
containing a table with different metadata compared to the table created on the
target side based on the PG metadata (as it uses ysql_dump).

Currently we try to match the source table and the target table by comparing
their schemas. Since they may mismatch, we can no longer match the tables.
Technically the source side does contain the YsqlDdlTxnVerifierState which
means the snapshot contains both the current and previous schemas. However
it is possible that we find a match for both the current and previous schemas.

Example: Say rename failed for a table because the name was already in use.
It could be possible for the rename to have succeeded on the DocDB side, leading
to 2 tables with the same name on YB-Master. If backup starts at this point, the
snapshot would contain two tables with the same name. When we iterate through
the tables in the snapshot, we would see this table with the transaction
verification state with 2 schemas. However we would find a table that matches
the current schema (the other table with the same name) and a table that matches
the previous schema (the old table before rename). We wouldn't have a simple way
of finding out which table is the matching table.

**Fix**
Fail backup if we find tables in the snapshot with YsqlDdlTxnVerifierState on
them.

**Other minor fixes**
The new tests caught the following problems, so their one-line fixes have been
sneaked into this diff

1.While matching schemas DDL Atomicity should handle the presence of column
**ybidxbasectid** which is a DocDB only column used in UNIQUE indexes
2.Missed a continue statement in YsqlTransactionDdl::MatchPgDocDBSchemaColumns
which was leading to issues in the happy path when ALTER TABLE DROP COLUMN
was successful. It was not caught earlier as existing tests only dropped the
last column in the schema.

**API/Gflag changes**
1.GetTableSchema RPC now additionally returns any YsqlDdlTxnVerifierState for
this table.  Currently this is only used by tests to wait for DDL verification
to end.
2.**FLAGS_retry_if_ddl_txn_verification_pending** is used to indicate whether
CatalogManager::AlterTable and CatalogManager::DeleteTable should best-effort
wait for any existing YsqlDdlTxnVerifierState on the table to be resolved before
throwing an error  (before this patch it threw error immediately). This flag
helps the case when sql scripts (such as running output of ysql_dump/PG regress
tests) run 2 DDLs on a table in quick succession. In this case, without this
best-effort wait, we might fail the sequentially occurring second DDL simply
because DDL verification has not ended. It is not easy for customers to
introduce retries in such DDL scripts, so this best effort check should be
useful. This fix is required in this diff as it is needed for backup tests to
finish successfully (as they run ysql_dump)
Jira: DB-2996

Test Plan:
Jenkins: urgent

./yb_build.sh debug --cxx-test yb-backup-cross-feature-test --gtest_filter "*YBDdlAtomicityBackupTest*"
./yb_build.sh debug --cxx-test pg_ddl_atomicity-test

I made some proto changes to YsqlDdlTxnVerifierState structure. It should
theoretically not affect upgrade as it does not change the actual number/type of
fields in the proto but I ran an upgrade test anyway.
1)Revert this diff. Enable DDL rollback but disable DDL verification. Run some
DDLs. This will lead to YB-Master writing SysTableEntryProto with the old format
into disk
2)Stop the cluster. Un-revert the diff and start the cluster with new changes
3)Verify that DDLs were verified and all future DML and DDL succeed with the
new proto changes

Reviewers: jhe, myang, loginov

Reviewed By: jhe

Subscribers: yql, ybase, bogdan

Tags: #jenkins-ready

Differential Revision: https://phorge.dev.yugabyte.com/D29521