[BACKPORT 2.20][#19605] YSQL: Fix test failures encountered upon enabling DDL Atomicity
Summary:
Original commit: a4e06b1bc86360b502d5ef4543250ff513fc6c1f / D29529
The following issues exists in DDL Atomicity:
1. If CREATE TABLE operation fails when the table is in PREPARING
state, then the DDL Atomicity code does not cleanup the table. This
is because, during rollback, a CHECK fires, expecting the table to either be in RUNNING/ALTERING
state. This check should be relaxed to include PREPARING state as well.

2. ALTER TABLE ALTER COLUMN TYPE with multiple sub-commands fails. This is because
it creates and drops tables multiple times within the transaction. If DDL Atomicity
code notices that a table is being created and deleted in the same transaction, it
does not postpone the drop until commit (this is because the table will be dropped
whether the transaction aborts or commits). However in case of such multi-rewrite alter,
the DROP operation causes all transactions operating on the table to Abort, and
causes the parent transaction to abort. Therefore, this optimisation should be removed,
and the DROP should be postponed to the end of commit.

3. When Materialized views are looked up in the PG schema from DocDB, they should
be looked up using the relfilenode and not the table oid.

4. DDL Atomicity code has an implicit assumption that the DocDB columns are sorted
by the "order" field. This is not always true, especially in the case where the
Primary key contains the columns in order different from the order of creation.
Jira: DB-8385

Test Plan:
Jenkins: urgent
./yb_build.sh --cxx-test pg_ddl_atomicity-test

Reviewers: fizaa, myang

Reviewed By: fizaa

Subscribers: bogdan, ybase, yql, jason

Tags: #jenkins-ready

Differential Revision: https://phorge.dev.yugabyte.com/D29554