[BACKPORT 2.20][PLAT-11196] Set the provider reference in image bundle as part of provider edit
Summary:
We were missing the provider reference on the imagebundle as part of provider edit, which
was causing failure in region add.
While the original issue is fixed as part of
https://github.com/yugabyte/yugabyte-db/blame/master/managed/src/main/java/com/yugabyte/yw/commissioner/tasks/subtasks/cloud/CloudImageBundleSetup.java#L237,
we don't want to backport the entire diff, thus the specific fix.

Test Plan: Manually tested region add in the provider.

Reviewers: vbansal

Reviewed By: vbansal

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D29891