[BACKPORT 2.20][PLAT-11216] Backup script failing on CentOS 7
Summary:
Original commit: 26133675b8bf83b026fa50d5b774be792033e357 / D29944
Restore was failing on prometheus directories because centos was taking single quotes as a literal. This diff removes the quotes so it treats prom_snapshot as a variable.

This was a regression from https://phorge.dev.yugabyte.com/D29466 which is in 2.18.4 and 2.20.0

Test Plan: Backup restore on ubuntu and centos 7 wih change, ensure prometheus dir are moved properly.

Reviewers: dshubin, sanketh

Reviewed By: dshubin

Subscribers: yugaware

Tags: #jenkins-ready

Differential Revision: https://phorge.dev.yugabyte.com/D29966