[BACKPORT 2.20][PLAT-10930] Permissions issue when creating yugabyte bash_profile
Summary:
Original commit: 319a09499a85031804d959782b20081822c87252 / D29614
On certain OS's when we create a user the bash_profile is not automatically created for that user.
Thus, when Ansible creates it, because we don't explicitly set the owner/group it is owned by root
and then throws permission errors whenever yugabyte user actually tries to read/execute the file.
This diff explicitly sets the ownership of the right files in case they are created.

Test Plan: universe creation on SLES

Reviewers: amalyshev, nbhatia

Reviewed By: nbhatia

Subscribers: yugaware

Tags: #jenkins-ready

Differential Revision: https://phorge.dev.yugabyte.com/D30284