[BACKPORT 2.20][PLAT-11352][xCluster] Do not pass materialized views when setting up replication
Summary:
This diff makes change to return an error when the user send a request to set up replication for tables that is not supported for xCluster replication.

Also, it adds a query string to the get tables api called onlySupportedForXCluster which returns only relation types there are supported to be in xCluster replication. It is used in the UI xCluster components to show only supported tables.

Original commit: ba77fc95c35544a38158b4481b2e2818a71bd046 / D30204

Test Plan:
- Made sure an error is returned if the tables list include an materialized view for both xCluster and DR APIs.
- Made sure if you don't pass the materialized view uuid, the API call to create an xCluster config and a DR config succeeds.
- Made sure user can only add supported tables to an existing config.
- Made sure UI shows the supported tables list only.

Reviewers: #yba-api-review!, cwang, jmak, sanketh

Reviewed By: cwang

Subscribers: yugaware

Tags: #jenkins-ready

Differential Revision: https://phorge.dev.yugabyte.com/D30363